Create reusable mairie components
